refactor agents project
