+++ auto-content-creator/agents/src/logger.ts
@@ -1,4 +1,5 @@
 import winston from 'winston';
+import { config } from './config';
 
 const logger = winston.createLogger({
   level: 'info',
@@ -8,14 +9,14 @@ const logger = winston.createLogger({
     winston.format.splat(),
     winston.format.json()
   ),
-  defaultMeta: { service: 'writer-agent' },
+  defaultMeta: { service: 'writer-agent', environment: config.environment },
   transports: [
     new winston.transports.File({ filename: 'error.log', level: 'error' }),
     new winston.transports.File({ filename: 'combined.log' }),
   ],
 });
 
-if (process.env.NODE_ENV !== 'production') {
+if (config.environment !== 'production') {
   logger.add(
     new winston.transports.Console({
       format: winston.format.combine(winston.format.colorize(), winston.format.simple()),
